add: implement solutions for problems "B. Transfusion" and "A. k-th divisor" in Codeforces Round 991 with test cases and configurations

+++ All Contest/Codeforces Round 991 (Div. 3)/B_Transfusion.cpp
@@ -13,8 +13,8 @@ int main(){
      int t; cin>>t;
      while (t--)
      {
-        int n;cin>>n;
-        int sum_o = 0, sum_e = 0;
+        ll n;cin>>n;
+        ll sum_o = 0, sum_e = 0;
         for (int i = 1; i <= n; i++)
         {
             int x;cin>>x;
